Use the MSHW ID to load firmware
The ID for IPTS is defined by a device with the ACPI path \\_SB.TSML. This
device seems to exist on all surface models, and does nothing except returning
that MSHW ID as it's _HID.

```
Device (TSML)
{
	Method (_HID, 0, NotSerialized)  // _HID: Hardware ID
	{
		Return ("MSHW0079")
	}
}
```

(this is from a Surface Laptop DSDT I found online. I checked another other
DSDT (SB1), and my own one via. sysfs, and all of them report the same path)

Since there is no way to grab the MSHW ID from the driver directly (as it uses
the MEI bus), we have to search for this device, and query it's _HID property.
This gives us the MSHW ID, and we prefix the firmware path with that, instead
of the model name / SKU.

This is a bit more hacky than the approach with model name and SKU, since it
depends on hardcoded ACPI paths, but it results in a more generic firmware
structure (no need to duplicate the files for models who expose the same _HID).

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>